+++ scan.h
@@ -1,7 +1,6 @@
 #ifndef SCAN_H
 #define SCAN_H
 
-#include <stddef.h>
 #include <stdbool.h>
 #include <ctype.h>
 
@@ -28,7 +27,7 @@ static inline bool isOperatorCharacter(char c) {
 
 const char* getFirstLexeme(const char* input);
 const char* getNextLexeme(const char* lastLexeme);
-size_t getLexemeLength(const char* lexeme);
+unsigned int getLexemeLength(const char* lexeme);
 bool isSameLexeme(const char* a, const char* b);
 
 #endif
